separate settings menu into settings and feature flags. Also puts export and tz in settings menu
