+++ openbb_terminal/timezone.openbb
@@ -1,1 +1,1 @@
-US/Eastern
+America/New_York
